fix: split embed view and schema for web worker compatibility (#7391)
BS-621
