+++ packages/blocks/src/embed-figma-block/embed-figma-spec.ts
@@ -1,31 +1,13 @@
+import type { BlockSpec } from '@blocksuite/block-std';
 import { literal } from 'lit/static-html.js';
 
-import { createEmbedBlock } from '../_common/embed-block-helper/index.js';
-import {
-  type EmbedFigmaBlockProps,
-  EmbedFigmaModel,
-  EmbedFigmaStyles,
-} from './embed-figma-model.js';
+import { EmbedFigmaBlockSchema } from './embed-figma-schema.js';
 import { EmbedFigmaBlockService } from './embed-figma-service.js';
 
-const defaultEmbedFigmaProps: EmbedFigmaBlockProps = {
-  style: EmbedFigmaStyles[0],
-  url: '',
-  caption: null,
-
-  title: null,
-  description: null,
-};
-
-export const EmbedFigmaBlockSpec = createEmbedBlock({
-  schema: {
-    name: 'figma',
-    version: 1,
-    toModel: () => new EmbedFigmaModel(),
-    props: (): EmbedFigmaBlockProps => defaultEmbedFigmaProps,
-  },
+export const EmbedFigmaBlockSpec: BlockSpec = {
+  schema: EmbedFigmaBlockSchema,
   view: {
     component: literal`affine-embed-figma-block`,
   },
   service: EmbedFigmaBlockService,
-});
+};
