+++ packages/blocks/src/embed-github-block/embed-github-schema.ts
@@ -1,0 +1,31 @@
+import { createEmbedBlockSchema } from '../_common/embed-block-helper/helper.js';
+import {
+  type EmbedGithubBlockProps,
+  EmbedGithubModel,
+  EmbedGithubStyles,
+} from './embed-github-model.js';
+
+const defaultEmbedGithubProps: EmbedGithubBlockProps = {
+  style: EmbedGithubStyles[1],
+  owner: '',
+  repo: '',
+  githubType: 'issue',
+  githubId: '',
+  url: '',
+  caption: null,
+
+  image: null,
+  status: null,
+  statusReason: null,
+  title: null,
+  description: null,
+  createdAt: null,
+  assignees: null,
+};
+
+export const EmbedGithubBlockSchema = createEmbedBlockSchema({
+  name: 'github',
+  version: 1,
+  toModel: () => new EmbedGithubModel(),
+  props: (): EmbedGithubBlockProps => defaultEmbedGithubProps,
+});
